First split
