+++ high_level/helpers.py
@@ -3,19 +3,21 @@ from pyrsistent import PSet, PMap, PVector
 from data_structures import *
 from formal_verification_annotations import *
 from pythonic_code_generic import *
+from common import *
+from bft import *
 from stubs import *
 
 
 def get_slot_from_time(time: int, node_state: NodeState) -> int:
     """
-    It calculates the slot number based on a given `time` and the node's configuration settings. 
+    It calculates the slot number based on a given `time` and the node's configuration settings.
     """
     return time // (4 * node_state.configuration.delta)
 
 
 def get_phase_from_time(time: int, node_state: NodeState) -> NodePhase:
     """
-    It calculates the phase within a slot based on a given `time` and the node's configuration settings. 
+    It calculates the phase within a slot based on a given `time` and the node's configuration settings.
     """
     time_in_slot = time % (4 * node_state.configuration.delta)
 
@@ -29,77 +31,9 @@ def get_phase_from_time(time: int, node_state: NodeState) -> NodePhase:
         return NodePhase.PROPOSE
 
 
-def genesis_checkpoint(node_state: NodeState) -> Checkpoint:
-    """
-    It defines the genesis block.
-    """
-    return Checkpoint(
-        block_hash=block_hash(node_state.configuration.genesis),
-        chkp_slot=0,
-        block_slot=0
-    )
-
-
-def has_block_hash(block_hash: Hash, node_state: NodeState) -> bool:
-    """
-    It checks if a given `block_hash` is present within a `node_state`.
-    """
-    return pmap_has(node_state.view_blocks, block_hash)
-
-
-def get_block_from_hash(block_hash: Hash, node_state: NodeState) -> Block:
-    """
-    It retrieves the block associated to a `block_hash`.
-    """
-    Requires(has_block_hash(block_hash, node_state))
-    return pmap_get(node_state.view_blocks, block_hash)
-
-
-def has_parent(block: Block, node_state: NodeState) -> bool:
-    """
-    It checks whether a `block` has a parent.
-    """
-    return has_block_hash(block.parent_hash, node_state)
-
-
-def get_parent(block: Block, node_state: NodeState) -> Block:
-    """
-    It retrieves the parent of a given `block`.
-    """
-    Requires(has_parent(block, node_state))
-    return get_block_from_hash(block.parent_hash, node_state)
-
-
-def get_all_blocks(node_state: NodeState) -> PSet[Block]:
-    """
-    It retrieves all the blocks in a `node_state`.
-    """
-    return pmap_values(node_state.view_blocks)
-
-
-def is_validator(node: NodeIdentity, validatorBalances: ValidatorBalances) -> bool:
-    """
-    It checks whether a `node` is a validator.
-    """
-    return pmap_has(validatorBalances, node)
-
-
-def is_complete_chain(block: Block, node_state: NodeState) -> bool:
-    """
-    It checks if a given `block` is part of a complete chain of blocks that reaches back to the genesis block `node_state.configuration.genesis`
-    within a `node_state`.
-    """
-    if block == node_state.configuration.genesis:
-        return True
-    elif not has_parent(block, node_state):
-        return False
-    else:
-        return is_complete_chain(get_parent(block, node_state), node_state)
-
-
 def get_blockchain(block: Block, node_state: NodeState) -> PVector[Block]:
     """
-    It constructs a blockchain from a given `block` back to the genesis block, 
+    It constructs a blockchain from a given `block` back to the genesis block,
     assuming the given `block` is part of a complete chain.
     """
     Requires(is_complete_chain(block, node_state))
@@ -112,107 +46,6 @@ def get_blockchain(block: Block, node_state: NodeState) -> PVector[Block]:
         )
 
 
-def is_ancestor_descendant_relationship(ancestor: Block, descendant: Block, node_state: NodeState) -> bool:
-    """
-    It determines whether there is an ancestor-descendant relationship between two blocks.
-    """
-    if ancestor == descendant:
-        return True
-    elif descendant == node_state.configuration.genesis:
-        return False
-    else:
-        return (
-            has_parent(descendant, node_state) and
-            is_ancestor_descendant_relationship(
-                ancestor,
-                get_parent(descendant, node_state),
-                node_state
-            )
-        )
-
-
-def validator_set_weight(validators: PSet[NodeIdentity], validatorBalances: ValidatorBalances) -> int:
-    """
-    It calculates the total weight (or sum of `validatorBalances`) of a specified set of `validators` within a blockchain.
-    """
-    return pset_sum(
-        pset_map(
-            lambda v: pmap_get(validatorBalances, v),
-            pset_intersection(
-                pmap_keys(validatorBalances),
-                validators
-            )
-        )
-    )
-
-
-def get_set_FFG_targets(votes: PSet[SignedVoteMessage]) -> PSet[Checkpoint]:
-    """
-    It extracts a set of ffg targets from a set of `votes`.
-    """
-    return pset_map(
-        lambda vote: vote.message.ffg_target,
-        votes
-    )
-
-
-def is_FFG_vote_in_support_of_checkpoint_justification(vote: SignedVoteMessage, checkpoint: Checkpoint, node_state: NodeState) -> bool:
-    """
-    It determines whether a given `vote` supports the justification of a specified `checkpoint`.
-    """
-    return (
-        valid_vote(vote, node_state) and
-        vote.message.ffg_target.chkp_slot == checkpoint.chkp_slot and
-        is_ancestor_descendant_relationship(
-            get_block_from_hash(checkpoint.block_hash, node_state),
-            get_block_from_hash(vote.message.ffg_target.block_hash, node_state),
-            node_state) and
-        is_ancestor_descendant_relationship(
-            get_block_from_hash(vote.message.ffg_source.block_hash, node_state),
-            get_block_from_hash(checkpoint.block_hash, node_state),
-            node_state) and
-        is_justified_checkpoint(vote.message.ffg_source, node_state)
-    )
-
-
-def filter_out_FFG_votes_not_in_FFG_support_of_checkpoint_justification(votes: PSet[SignedVoteMessage], checkpoint: Checkpoint, node_state: NodeState) -> PSet[SignedVoteMessage]:
-    """
-    It filters out ffg votes that do not support the justification of a specified `checkpoint`.
-    """
-    return pset_filter(lambda vote: is_FFG_vote_in_support_of_checkpoint_justification(vote, checkpoint, node_state), votes)
-
-
-def get_validators_in_FFG_support_of_checkpoint_justification(votes: PSet[SignedVoteMessage], checkpoint: Checkpoint, node_state: NodeState) -> PSet[NodeIdentity]:
-    """
-    It identifies and returns the set of validators that have cast `votes` in support of the justification of a specified `checkpoint`.
-    """
-    return pset_map(
-        lambda vote: vote.sender,
-        filter_out_FFG_votes_not_in_FFG_support_of_checkpoint_justification(votes, checkpoint, node_state)
-    )
-
-
-def is_justified_checkpoint(checkpoint: Checkpoint, node_state: NodeState) -> bool:
-    """
-    It checks whether a `checkpoint` if justified, specifically a `checkpoint` is justified if at least 
-    two-thirds of the total validator set weight is in support. This is evaluated by checking if 
-    `FFG_support_weight * 3 >= tot_validator_set_weight * 2`.
-    """
-
-    if checkpoint == genesis_checkpoint(node_state):
-        return True
-    else:
-        if not has_block_hash(checkpoint.block_hash, node_state) or not is_complete_chain(get_block_from_hash(checkpoint.block_hash, node_state), node_state):
-            return False
-
-        validatorBalances = get_validator_set_for_slot(get_block_from_hash(checkpoint.block_hash, node_state), checkpoint.block_slot, node_state)
-
-        FFG_support_weight = validator_set_weight(get_validators_in_FFG_support_of_checkpoint_justification(node_state.view_votes, checkpoint, node_state), validatorBalances)
-        tot_validator_set_weight = validator_set_weight(pmap_keys(validatorBalances), validatorBalances)
-
-        return FFG_support_weight * 3 >= tot_validator_set_weight * 2
-
-
 def filter_out_non_justified_checkpoint(checkpoints: PSet[Checkpoint], node_state: NodeState) -> PSet[Checkpoint]:
     """
     It filters out `checkpoints` that are not justified.
@@ -222,7 +55,7 @@ def filter_out_non_justified_checkpoint(checkpoints: PSet[Checkpoint], node_stat
 
 def get_justified_checkpoints(node_state: NodeState) -> PSet[Checkpoint]:
     """
-    It retrieves all the justified checkpoints from a `note_state`. First it extracts all ffg target checkpoints from 
+    It retrieves all the justified checkpoints from a `note_state`. First it extracts all ffg target checkpoints from
     the set of votes in the `node_state`. Then it filter out these checkpoints to keep only those that are justified.
     The `genesis_checkpoint` is automatically considered justified.
     """
@@ -239,80 +72,6 @@ def get_greatest_justified_checkpoint(node_state: NodeState) -> Checkpoint:
     return pset_max(
         get_justified_checkpoints(node_state),
         lambda c: (c.chkp_slot, c.block_slot)
-    )
-
-
-def is_FFG_vote_linking_to_a_checkpoint_in_next_slot(vote: SignedVoteMessage, checkpoint: Checkpoint, node_state: NodeState) -> bool:
-    """
-    It evaluates whether a given `vote` represents a link from a specified `checkpoint` to a checkpoint in the immediately following slot.
-    """
-    return (
-        valid_vote(vote, node_state) and
-        vote.message.ffg_source == checkpoint and
-        vote.message.ffg_target.chkp_slot == checkpoint.chkp_slot + 1
-    )
-
-
-def filter_out_FFG_vote_not_linking_to_a_checkpoint_in_next_slot(checkpoint: Checkpoint, node_state: NodeState) -> PSet[SignedVoteMessage]:  
-    """
-    It filters and retains only those votes from a `node_state` that are linking to a `checkpoint` in the next slot.
-    """
-    return pset_filter(lambda vote: is_FFG_vote_linking_to_a_checkpoint_in_next_slot(vote, checkpoint, node_state), node_state.view_votes)
-
-
-def get_validators_in_FFG_votes_linking_to_a_checkpoint_in_next_slot(checkpoint: Checkpoint, node_state) -> PSet[NodeIdentity]:
-    """
-    It retrieves the identities of validators who have cast ffg votes that support linking a specified `checkpoint` to its immediate successor.
-    """
-    return pset_map(
-        lambda vote: vote.sender,
-        filter_out_FFG_vote_not_linking_to_a_checkpoint_in_next_slot(checkpoint, node_state)
-    )
-
-
-def is_finalized_checkpoint(checkpoint: Checkpoint, node_state: NodeState) -> bool:
-    """
-    It evaluates whether a given `checkpoint` has been finalized. A `checkpoint` is considered finalized if it is justified and 
-    if at least two-thirds of the total validator set's weight supports the transition from this `checkpoint` to the next. Specifically, it first checks if the `checkpoint` is justified using 
-    `is_justified_checkpoint(checkpoint, node_state)`. Then it retrieves the set of validators and their balances for the slot associated with the `checkpoint`. 
-    This is done through `get_validator_set_for_slot`. Then it calculates the total weight (stake) of validators who have cast votes 
-    linking the `checkpoint` to the next slot, using `get_validators_in_FFG_votes_linking_to_a_checkpoint_in_next_slot` to identify these validators
-    and `validator_set_weight` to sum their stakes. Finally it checks if `FFG_support_weight * 3 >= tot_validator_set_weight * 2` to finalize `checkpoint`.
-    """
-    if not is_justified_checkpoint(checkpoint, node_state):
-        return False
-
-    validatorBalances = get_validator_set_for_slot(get_block_from_hash(checkpoint.block_hash, node_state), checkpoint.block_slot, node_state)
-    FFG_support_weight = validator_set_weight(get_validators_in_FFG_votes_linking_to_a_checkpoint_in_next_slot(checkpoint, node_state), validatorBalances)
-    tot_validator_set_weight = validator_set_weight(pmap_keys(validatorBalances), validatorBalances)
-
-    return FFG_support_weight * 3 >= tot_validator_set_weight * 2
-
-
-def filter_out_non_finalized_checkpoint(checkpoints: PSet[Checkpoint], node_state: NodeState) -> PSet[Checkpoint]:
-    """
-    It filters out non finalized `checkpoints` from a `node_state`.
-    """
-    return pset_filter(lambda checkpoint: is_finalized_checkpoint(checkpoint, node_state), checkpoints)
-
-
-def get_finalized_checkpoints(node_state: NodeState) -> PSet[Checkpoint]:
-    """
-    It retrieves from `node_state` all the checkpoints that have been finalized.
-    """
-    return pset_add(
-        filter_out_non_finalized_checkpoint(get_set_FFG_targets(node_state.view_votes), node_state),
-        genesis_checkpoint(node_state)
-    )
-
-
-def get_greatest_finalized_checkpoint(node_state: NodeState) -> Checkpoint:
-    """
-    It returns the greatest finalized checkpoint from a `node_state`.
-    """
-    return pset_max(
-        get_finalized_checkpoints(node_state),
-        lambda c: c.chkp_slot
     )
 
 
@@ -367,7 +126,7 @@ def filter_out_GHOST_votes_not_for_blocks_in_blockchain(votes: PSet[SignedVoteMe
 
 def filter_out_GHOST_votes_for_blocks_in_blockchain(votes: PSet[SignedVoteMessage], blockchainHead: Block, node_state: NodeState) -> PSet[SignedVoteMessage]:
     """
-    It filter through a set of `votes`, excluding those that are for blocks within the blockchain leading up to and including a specified `blockchainHead`. 
+    It filter through a set of `votes`, excluding those that are for blocks within the blockchain leading up to and including a specified `blockchainHead`.
     """
     return pset_filter(
         lambda vote: not is_GHOST_vote_for_block_in_blockchain(vote, blockchainHead, node_state),
@@ -404,8 +163,8 @@ def filter_out_non_LMD_GHOST_votes(votes: PSet[SignedVoteMessage]) -> PSet[Signe
 
 def is_equivocating_GHOST_vote(vote: SignedVoteMessage, node_state: NodeState) -> bool:
     """
-    It checks if the given `vote` is part of an equivocation by comparing it against all other `vote`s from the same sender 
-    for the same slot but with different block hashes. If such a `vote` exists, the validator is considered to have equivocated, 
+    It checks if the given `vote` is part of an equivocation by comparing it against all other `vote`s from the same sender
+    for the same slot but with different block hashes. If such a `vote` exists, the validator is considered to have equivocated,
     violating the protocol's rules.
     """
     return not pset_is_empty(
@@ -420,8 +179,8 @@ def is_equivocating_GHOST_vote(vote: SignedVoteMessage, node_state: NodeState) -
 
 
 def filter_out_GHOST_equivocating_votes(votes: PSet[SignedVoteMessage], node_state: NodeState) -> PSet[SignedVoteMessage]:
-    """ 
-    It filters out from `votes` all the equivocating votes. 
+    """
+    It filters out from `votes` all the equivocating votes.
     """
     return pset_filter(
         lambda vote: not is_equivocating_GHOST_vote(vote, node_state),
@@ -433,37 +192,24 @@ def valid_vote(vote: SignedVoteMessage, node_state: NodeState) -> bool:
     """
     A vote is valid if:
     - it has a valid signature;
+    - it is a valid FFG vote
     - the block hash associated with the voted head block exists within a validator's view of blocks;
     - the head block associated with the vote is part of a complete chain that leads back to the genesis block within a validator's state;
     - the sender is a validator;
     - `vote.message.ffg_source.block_hash` is an ancestor of `vote.message.ffg_target.block_hash`;
-    - `vote.message.ffg_target.block_hash` is an ancestor of `vote.message.head_hash`;
-    - the checkpoint slot of `vote.message.ffg_source` is strictly less than checkpoint slot of `vote.message.ffg_target`;   
-    - the block associated with `vote.message.ffg_source.block_hash` has a slot number that matches the slot number specified in the same vote message;
-    - the block associated with `vote.message.ffg_target.block_hash` has a slot number that matches the slot number specified in the same vote message;
-    - the block hash associated the source exists within a validator's view of blocks; and
-    - the block hash associated the target exists within a validator's view of blocks.
     """
     return (
         verify_vote_signature(vote) and
+        valid_FFG_vote(vote, node_state) and
         has_block_hash(vote.message.head_hash, node_state) and
         is_complete_chain(get_block_from_hash(vote.message.head_hash, node_state), node_state) and
         is_validator(
             vote.sender,
             get_validator_set_for_slot(get_block_from_hash(vote.message.head_hash, node_state), vote.message.slot, node_state)) and
         is_ancestor_descendant_relationship(
-            get_block_from_hash(vote.message.ffg_source.block_hash, node_state),
-            get_block_from_hash(vote.message.ffg_target.block_hash, node_state),
-            node_state) and
-        is_ancestor_descendant_relationship(
             get_block_from_hash(vote.message.ffg_target.block_hash, node_state),
             get_block_from_hash(vote.message.head_hash, node_state),
-            node_state) and
-        vote.message.ffg_source.chkp_slot < vote.message.ffg_target.chkp_slot and
-        has_block_hash(vote.message.ffg_source.block_hash, node_state) and
-        get_block_from_hash(vote.message.ffg_source.block_hash, node_state).slot == vote.message.ffg_source.block_slot and
-        has_block_hash(vote.message.ffg_target.block_hash, node_state) and
-        get_block_from_hash(vote.message.ffg_target.block_hash, node_state).slot == vote.message.ffg_target.block_slot
+            node_state)
     )
 
 
@@ -543,9 +289,9 @@ def get_votes_to_include_in_propose_message_view(node_state: NodeState) -> PVect
 
 def get_GHOST_weight(block: Block, votes: PSet[SignedVoteMessage], node_state: NodeState, validatorBalances: ValidatorBalances) -> int:
     """
-    The GHOST weight of a `block` is determined by the total stake supporting the branch that ends with this `block` as its tip. 
+    The GHOST weight of a `block` is determined by the total stake supporting the branch that ends with this `block` as its tip.
     Validators vote with associated stakes, and the collective stake behind these votes establishes the block's GHOST weight.
-    """    
+    """
     return pset_sum(
         pset_map(
             lambda vote: validatorBalances[vote.sender],
@@ -566,7 +312,7 @@ def get_GHOST_weight(block: Block, votes: PSet[SignedVoteMessage], node_state: N
 def get_children(block: Block, node_state: NodeState) -> PSet[Block]:
     """
     Returns all the children of a given `block`.
-    """ 
+    """
     return pset_filter(
         lambda b: b.parent_hash == block_hash(block),
         get_all_blocks(node_state)
@@ -575,8 +321,8 @@ def get_children(block: Block, node_state: NodeState) -> PSet[Block]:
 
 def find_head_from(block: Block, votes: PSet[SignedVoteMessage], node_state: NodeState, validatorBalances: ValidatorBalances) -> Block:
     """
-    For a given `block`, it uses `get_GHOST_weight` to determine the chain's tip with the largest associated total stake.    
-    """ 
+    For a given `block`, it uses `get_GHOST_weight` to determine the chain's tip with the largest associated total stake.
+    """
     children = get_children(block, node_state)
 
     if len(children) == 0:
@@ -592,7 +338,7 @@ def find_head_from(block: Block, votes: PSet[SignedVoteMessage], node_state: Nod
 
 def get_head(node_state: NodeState) -> Block:
     """
-    It defines the fork-choice function. It starts from the greatest justified checkpoint, it considers 
+    It defines the fork-choice function. It starts from the greatest justified checkpoint, it considers
     the latest (non equivocating) votes cast by validators that are not older than `node_state.current_slot` - `node_state.configuration.eta` slots,
     and it outputs the head of the canonical chain with the largest associated total stake among such `relevant_votes`.
     """
@@ -629,10 +375,10 @@ def get_head(node_state: NodeState) -> Block:
 
 def execute_view_merge(node_state: NodeState) -> NodeState:
     """
-    It merges a validator's buffer with its local view, specifically merging the buffer of blocks `node_state.buffer_blocks` 
-    into the local view of blocks `node_state.view_blocks` and the buffer of votes `node_state.buffer_votes` into the 
+    It merges a validator's buffer with its local view, specifically merging the buffer of blocks `node_state.buffer_blocks`
+    into the local view of blocks `node_state.view_blocks` and the buffer of votes `node_state.buffer_votes` into the
     local view of votes `node_state.view_votes`.
-    """ 
+    """
     node_state = node_state.set(blocks=pmap_merge(node_state.view_blocks, node_state.buffer_blocks))
     node_state = node_state.set(view_vote=pset_merge(
         pset_merge(
@@ -649,7 +395,7 @@ def execute_view_merge(node_state: NodeState) -> NodeState:
 def get_block_k_deep(blockHead: Block, k: int, node_state: NodeState) -> Block:
     """
     It identifies the block that is `k` blocks back from the tip of the canonical chain, or the genesis block `node_state.configuration.genesis`.
-    """ 
+    """
     Requires(is_complete_chain(blockHead, node_state))
     if k <= 0 or blockHead == node_state.configuration.genesis:
         return blockHead
